requirements.txt
crewai[google-genai]>=0.80.0
python-dotenv>=1.0.0
pydantic>=2.0.0
